requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
attrs==22.1.0
    # via pytest
black==22.10.0
    # via -r requirements.in
build==0.9.0
    # via pip-tools
click==8.1.3
    # via
    #   black
    #   pip-tools
exceptiongroup==1.0.4
    # via pytest
iniconfig==1.1.1
    # via pytest
mypy-extensions==0.4.3
    # via black
packaging==21.3
    # via
    #   build
    #   pytest
pathspec==0.10.2
    # via black
pep517==0.13.0
    # via build
pip-tools==6.10.0
    # via -r requirements.in
platformdirs==2.5.4
    # via black
pluggy==1.0.0
    # via pytest
protobuf==3.20.3
    # via -r requirements.in
pyparsing==3.0.9
    # via packaging
pytest==7.2.0
    # via -r requirements.in
pyzmq==24.0.1
    # via zmq
tomli==2.0.1
    # via
    #   black
    #   build
    #   pytest
websocket-client==1.4.2
    # via -r requirements.in
wheel==0.38.4
    # via pip-tools
zmq==0.0.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
